etl_1c/requirements.txt
greenlet==2.0.1
numpy==1.23.5
pandas==1.5.2
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.46
xlrd==2.0.1
